handle already submitted claims
